TahlilchiAI/tahlilchi-rag/requirements.txt
# FastAPI and web server
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
greenlet==3.1.1
alembic==1.13.3

# Configuration and environment
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1

# File handling and authentication
python-multipart==0.0.20
email-validator==2.2.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (for external API calls)
httpx==0.28.1

# Background task processing and rate limiting
celery==5.4.0
redis==5.2.1

# Monitoring and observability
prometheus-client==0.21.1
python-json-logger==3.2.1

# Vector search and retrieval
qdrant-client==1.11.1
sentence-transformers==3.3.1
torch==2.9.1
rank-bm25==0.2.2

# Document parsing
pypdf==5.1.0
python-docx==1.1.2
langdetect==1.0.9

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
faker==33.1.0

# Development tools
black==25.11.0
ruff==0.14.5
mypy==1.18.2